选择空间文件夹

+++ Picture/picture-picture/src/main/java/com/lz/picture/model/vo/pictureInfo/PictureInfoVo.java
@@ -71,7 +71,7 @@ public class PictureInfoVo implements Serializable {
      * 宽高比例
      */
     @Excel(name = "宽高比例")
-    private Long picScale;
+    private Double picScale;
 
     /**
      * 图片格式
